Create package ds3autogen.utils & move Helper from ds3autogen.java; refactor all references
