+++ ds3-autogen-java/build.gradle
@@ -1,4 +1,5 @@
 dependencies {
     compile project(':ds3-autogen-api')
+    compile project(':ds3-autogen-utils')
     testCompile project(':ds3-autogen')
 }
