+++ ds3-autogen-utils/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    compile project(':ds3-autogen-api')
+}
